+++ src/main/java/com/autotune/experimentManager/services/CreateExperimentTrial.java
@@ -18,8 +18,8 @@ package com.autotune.experimentManager.services;
 import com.autotune.common.experiments.ExperimentTrial;
 import com.autotune.experimentManager.core.ExperimentTrialHandler;
 import com.autotune.experimentManager.data.ExperimentDetailsMap;
-import com.autotune.experimentManager.data.access.ExperimentAccess;
-import com.autotune.experimentManager.data.access.ExperimentAccessImpl;
+import com.autotune.experimentManager.data.TrialInterface;
+import com.autotune.experimentManager.data.TrialInterfaceImpl;
 import com.autotune.experimentManager.utils.EMConstants;
 import com.google.gson.Gson;
 import org.slf4j.Logger;
@@ -68,9 +68,9 @@ public class CreateExperimentTrial extends HttpServlet {
             String inputData = request.getReader().lines().collect(Collectors.joining());
             ExperimentTrial[] experimentTrialArray = gson.fromJson(inputData, ExperimentTrial[].class);
             List<ExperimentTrial> experimentTrialList = Arrays.asList(experimentTrialArray);
-            ExperimentAccess experimentAccess = new ExperimentAccessImpl(existExperimentTrialMap);
-            experimentAccess.addExperiments(experimentTrialList);
-            if (null == experimentAccess.getErrorMessage()) {
+            TrialInterface trialInterface = new TrialInterfaceImpl(existExperimentTrialMap);
+            trialInterface.addExperiments(experimentTrialList);
+            if (null == trialInterface.getErrorMessage()) {
                 experimentTrialList.forEach(
                         (experimentTrial) -> {
                             LOGGER.debug("Experiment name {} started processing", experimentTrial.getExperimentName());
@@ -83,13 +83,13 @@ public class CreateExperimentTrial extends HttpServlet {
                         }
                 );
             } else {
-                response.sendError(experimentAccess.getHttpResponseCode(), experimentAccess.getErrorMessage());
+                response.sendError(trialInterface.getHttpResponseCode(), trialInterface.getErrorMessage());
             }
             response.setContentType(JSON_CONTENT_TYPE);
             response.setCharacterEncoding(CHARACTER_ENCODING);
             response.setStatus(HttpServletResponse.SC_CREATED);
             PrintWriter out = response.getWriter();
-            out.append(new Gson().toJson(experimentAccess.listExperiments()));
+            out.append(new Gson().toJson(trialInterface.listExperiments()));
             out.flush();
         } catch (Exception e) {
             LOGGER.error(e.toString());
